break out payment logic
